Add the regression test.
